+++ config/routes.rb
@@ -10,5 +10,9 @@ Rails.application.routes.draw do
   # get "service-worker" => "rails/pwa#service_worker", as: :pwa_service_worker
 
   # Defines the root path route ("/")
-  # root "posts#index"
+  root "legal#privacy"
+
+  get "/privacy", to: "legal#privacy"
+  get "/terms", to: "legal#terms"
+  get "/accessibility", to: "legal#accessibility"
 end